Add notification of Terms of Service in every signed-in demo session
As per the Legal's request, since users never accept Google's ToS while
using devices with Demo Accounts on retail stores, we need to display a
legal message to users who will be interacting with the devices.

In each demo mode signed-in session, a notification is displayed when
the user interacts with the device (and breaks the attract loop) for the
first time. The title, the message and the link are provided by the
Legal team.

Screenshot of the new notification:
https://screenshot.googleplex.com/6M9tGAB7UYwHu7e

BUG=404873865 TEST=unit test DemoModeIdleHandlerTest.*

Change-Id: I99b232845871c16c82532cdfeb4b2c1f8058df4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6619350
Reviewed-by: Xiqi Ruan <xiqiruan@chromium.org>
Reviewed-by: Andre Le <leandre@chromium.org>
Commit-Queue: Haifan Wang <wanghaifan@google.com>
Cr-Commit-Position: refs/heads/main@{#1477456}