Display message in menu when content filters are enabled
Display a "Content filters are on" entry in the menu when the local
content filters settings are enabled for the user on the device.
Clicking it opens the content filter help center article.

This change is protected by
PropagateDeviceContentFiltersToSupervisedUser.

NO_IFTTT=New markers introduced

Bug: 409238951, 426817307
Change-Id: I1efae24422ebe834eb95f6738dbdf8a709ad7de4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6656819
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Tomek Jurkiewicz <tju@google.com>
Cr-Commit-Position: refs/heads/main@{#1477463}