Put testing profiles in user data dir
This will fix warnings from profile manager:
"Failed to initialize prefs for a profile at invalid path", which
show up in test failures but aren't the cause of the failures.

Also adds logging to test_file_util's PathDeleter about the
files in the directory that it failed to delete, to help diagnose
test failures.

Bug: 410751413
Change-Id: I68086e7f7ec216d503a6bd8b3dcb11fa268eeb76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6654413
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: David Bienvenu <davidbienvenu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477473}