Cache an HMONITOR in each ScreenWinDisplay to optimize display lookups
This optimization is behind two flags that must both be enabled:
base::features::kReducePPMs flag which groups several performance
optimizations to measure their combined impact, and
display::features::kScreenWinDisplayLookupByHMONITOR which can be used
as an independent killswitch.

The flags change the implementation of display lookups:

 * Both enabled:
   [HWND, Rect, Point] -> HMONITOR Keyed Lookup
 * Either disabled (status quo):
   [HWND, Rect, Point] -> HMONITOR -> MONITORINFOEX -> ID lookup

ScreenWin::GetScreenDisplayVia is causing about 2.6% of all Chrome jank
on Windows, almost all of it in MonitorInfoFromHMONITOR and
DisplayIdFromMonitorInfo. (See http://screen/7m4rm7HhnYg5fgR.png .) This
is because GetScreenWinDisplayVia needs to convert an HMONITOR to
MONITORINFOEX, and then to a stable display ID, in order to look up the
ScreenWinDisplay by ID. The conversion has a lot of overhead and happens
very often.

According to
https://learn.microsoft.com/en-us/windows/win32/gdi/hmonitor-and-the-device-context,
"A physical display has the same HMONITOR as long as it is part of the
desktop. When a WM_DISPLAYCHANGE message is sent, any monitor may be
removed from the desktop and thus its HMONITOR becomes invalid or has
its settings changed." So the mapping of HMONITOR to ScreenWinDisplay
should only change at WM_DISPLAYCHANGE.

This adds an optional<HMONITOR> to ScreenWinDisplay, which is set to
nullopt on WM_DISPLAYCHANGE. (And initialized to nullopt for displays
that never had an HMONITOR, such as on headless.) In practice the list
of displays is regenerated with UpdateAllDisplaysAndNotify() after
WM_DISPLAYCHANGE, which assigns new HMONITOR's to each display, but if
any display remains in the list without a new HMONITOR its stale value
is cleared.

With the optimization enabled GetScreenDisplayVia looks up the
ScreenDisplayWin by HMONITOR and falls back to the
HMONITOR->MONITORINFOEX->stable ID slow path if none is found.

Note that ScreenWinHeadless doesn't use HMONITOR at all, so must still
look up ScreenDisplayWin by converting MONITORINFOEX to stable ID. This
is fast because ScreenWinHeadless uses synthesized display ID's,
avoiding the overhead of looking up HMONITOR properties.

Several new ScreenWin methods are made virtual, so that they can be
overridden in tests and for headless:

 * HMONITORFrom[ScreenPoint|ScreenRect|Window]: New methods, analogues
   of MonitorInfoFrom[ScreenPoint|ScreenRect|Window]. Overridden in
   tests to return fake HMONITORs, and in ScreenWinHeadless to do
   nothing.
 * MonitorInfoFromHMONITOR: Formerly a private function. Overridden in
   tests to map fake HMONITORs to monitor info, and in
   ScreenWinHeadless to do nothing.
 * MonitorInfoFrom[ScreenPoint|ScreenRect|Window]: Formerly overridden
   in tests and in ScreenWinHeadless. Now only overridden in
   ScreenWinHeadless. Tests use the production implementations, which
   call their overridden HMONITORFrom* and MonitorInfoFromHMONITOR.
 * GetScreenWinDisplayNearestScreen[Rect|Point]: Formerly non-virtual.
   Overridden in ScreenWinHeadless to skip the HMONITOR lookup.
   (For some reason GetScreenWinDisplayNearestHWND already had an
   override in ScreenWinHeadless.)
 * GetScreenWinDisplayForHMONITOR: New method that does the HMONITOR
   lookup. Overridden in ScreenWinHeadless to do nothing.

Bug: 424432184
Change-Id: I1a67ffffbde987d6b0e28054f4ce85bf765b1803
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6651137
Reviewed-by: Robert Liao <robliao@chromium.org>
Reviewed-by: Daniel Nicoara <dnicoara@chromium.org>
Commit-Queue: Joe Mason <joenotcharles@google.com>
Cr-Commit-Position: refs/heads/main@{#1477495}