Mitigate web-transport throttling based on the end-point's IP
This CL changes the HTTP3 WebTransport client to pass the server's IP
in case of failure during the handshake negotiation, as described in
the design document [1].

A new method is added to the WebTransport mojom interface of the
Network Service to pass a IPEndPoint instance to the browser process
before the UDP connection is created.

The blink's implementation of this interface will send an invalid
IPEndPoint instance to avoid handling sensitive information in the
renderer process.

The WebTransportThrottling class will receive the IPEndPoint instance
in the new argument of the Tracker::InHandshakeFailed function; it will
be used then to compute the appropriate penalty. A new method is added
specifically for this purpose.

[1]https://docs.google.com/document/d/1j2NJqLQyAHSBZArW6Iwg8uVId3NGQxnBV7b9LYbd8FM/edit#heading=h.numlu03ovx44

Bug: 40069954
Change-Id: Id59b195aee4e2b2a227f0a56878629a38d7ec8ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5756508
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Javier Fernandez <jfernandez@igalia.com>
Reviewed-by: Adam Rice <ricea@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477526}