Renderer navigation timeline part 2: add renderer-side traces and metrics
This CL uses the navigation_metrics_token introduced in
https://crrev.com/c/6514188 to record a new set of renderer-centric
navigation trace events, representing a timeline of how the
destination renderer process handles a navigation, starting from
process startup and progressing through view creation, proxy creation,
provisional frame creation, and finally CommitNavigation. The
browser-to-renderer IPCs used in this CL to create the new timeline
are CreateView, CreateRemoteChildren, CreateFrame, and
CommitNavigation. Note that no timelines are shown for other renderer
processes involved in the navigation, including the initiator process
or intermediate processes if a redirect occurs.

The approach is to introduce a new renderer-global
RendererNavigationMetricsManager object which keeps a map from
navigation metrics tokens to a per-navigation timeline object that
keeps timestamps for events pertaining to that navigation. Note that
there could be multiple CreateView and CreateRemoteChildren IPCs per
navigation, for example when setting up proxies in multiple pages on
the navigating frame's opener chain. Hence, there is a list of
begin/end timestamps for these IPCs, rather than a single pair. Once a
navigation finishes processing CommitNavigation (or its variants
CommitFailedNavigation/CommitSameDocumentNavigation), the
RendererNavigationMetricsManager is given a signal to record the
timeline, and it does this following precedent set in
https://crrev.com/c/6092101.

Handling aborted or canceled navigations is fairly complex. Ideally,
we'd clean up a navigation's timestamps when its NavigationClient
interface gets disconnected (which is how the renderer knows that
navigations are aborted), but sadly the frame/view/proxy creation IPCs
are not sent via NavigationClient and moreover, they are sent before
the NavigationClient is even set up in some cases (e.g.,
cross-process, browser-initiated navigations). Changing
NavigationClient to be set up sooner might work as a longer-term
solution, but would face its own challenges, since it's current
lifetime is tied to web compatibility, by defining when a
renderer-initiated navigation may be canceled via JS (see comments on
RenderFrameImpl::navigation_client_impl_). In the short term, this CL
takes a simple best-effort approach of cleaning up a navigation's
timestamps if the navigation hasn't committed after a timeout of 5
minutes (note that this covers at least 99.99%+ of navigations as
measured by Navigation.StartToCommit, which is capped at 3 minutes).

Due to https://crbug.com/405437928, overlapping navigations may not
show up properly nested; this will be investigated and fixed
separately, both for the existing browser timeline events and this
CL's renderer timeline events.

See the "Renderer-centric track" proposal in
https://docs.google.com/document/d/1WUTAA9mbsr__prHyYPFPWKqNQpX0x7HVjoYvqt_n_q0/edit?tab=t.0#heading=h.p50kjlrjeqlr
for more information.

This CL is not meant to be exhaustive but rather to introduce a
foundation with core trace events. Events for processing other IPCs
pertaining to navigations might be added to this timeline in the
future. Also, one known limitation is that renderer-initiated
same-document and synchronous about:blank navigations currently don't
show up on the renderer timeline, as they involve no IPCs from the
browser at all.

Sample traces:

NavigateABCToABD test:
https://ui.perfetto.dev/#!/?s=15ba048cb3d634da74323b39ad6a15a68d0beda1

Beforeunload handling in
RenderFrameHostImplWithOrWithoutLegacyBeforeUnloadBrowserTest.BeforeUnloadDialogInOOPIF:
https://ui.perfetto.dev/#!/?s=08f9af60c67258767dcf518955d7bdeec0ed86d6

Opener chains in PostMessageWithSubframeOnOpenerChain:
https://ui.perfetto.dev/#!/?s=e041901eb98b96d74f082a5944b597c186a75085
E.g., the last baz.com navigation has a length-2 opener chain and
creates three views: one for opener.opener, one for opener, and one
for itself.

Remote-to-local navigation in
SitePerProcessScrollAnchorTest.RemoteToLocalScrollAnchorRestore:
https://ui.perfetto.dev/#!/?s=5ce029b799ab6b34365e22128140d0700f7964aa
(The last navigation is remote-to-local and starts with CreateFrame
rather than CreateView.)

Browser-initiated same-document navigation in
NavigationBrowserTest.SameDocumentNavigation:
https://ui.perfetto.dev/#!/?s=39371e6b5b7c5f44dabdeb02986f8d46e121732d
(Second navigation is same-document and an example of how the commit
IPC is the navigation's first IPC, and no frames/views/proxies are
created.)

Bug: 415821826
Change-Id: Icf99bd35c5aa6b24996dd621a6614e225e3f6892
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6292796
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477530}