ozone/drm: Disable all CRTCs/connectors for tiled controller
Currently, ScreenManager::RemoveDisplayControllers() removes a
HardwareDisplayController from its mapping when a single CRTC is
removed from a HDC. In the tiled display case when multiple
CRTC/Connector pairs  may exist in a HDC, some enabled pairings may not
be disabled due to the HDC being removed so early.

This CL ensures that all CRTCs/controllers in a tiled HDC are part of
the disable commit request. Removing a tiled HDC like this is OK (and
even preferable) vs waiting for all CRTC/connectors to be removed, as
DrmGpuDisplayManager is going to create a new DrmDisplay and HDC if
there's a remaining tiled connector anyways; not erasing it just causes
confusion and the screen to freeze in occasion due to mismatched
controller mapping.

Bug: b:422857452
Test: ozone_unittests && manual testing
Change-Id: I014ef37e6d88313be9e16924340d19596bddc33f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6651636
Reviewed-by: Gil Dekel <gildekel@chromium.org>
Commit-Queue: Su Hong Koo <sukoo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477537}