[Watermark 2.0] Fix incorrect check for watermark command-line flags
The logic to check if the current channel is not STABLE or BETA was
incorrect. It used an || (OR) operator which would always evaluate to
true, meaning the command-line flags for watermark opacity were always
being checked, regardless of the channel.

This change corrects the logic to use && (AND), so the flags are only
checked on non-stable and non-beta channels as intended.

Bug: 426645402
Change-Id: Iff9a8745c25a84b79d6704e6d0b960041349d7c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6662822
Reviewed-by: Nasser Al-shawwa <alshawwa@chromium.org>
Commit-Queue: Adam Kassi Lahlou <adamkl@google.com>
Reviewed-by: Dominique Fauteux-Chapleau <domfc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477559}