[Settings][DBD] Fix subpage linkrows border-radius logic
Using existing CSS selectors like nth-child or nth-of-type does not work
given the dynamic visibility and different elements being used for the
linkout rows in the 'Other google data' subpage.

nth-of-type targets elements of the same type.

nth-child targets any elements in the dom, however, given that the
elements are shown conditionally, wrapping them in a dom-if template
(with restamp) still leaves behind an element with `display: none` which
the selector would interpret as a child.

This change determines the appropriate CSS class programmatically
instead depending on the visibility of the linkouts.

Bug: 397187800
Change-Id: I720f52a785cc7163b3c11681100200daf0a3c5f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6662278
Commit-Queue: Zaina Al-Mashni <zalmashni@google.com>
Reviewed-by: Rainhard Findling <rainhard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477564}