[Multiple Requests] Integrate with virtual card enrollment flow (4)
Have CCAM call VCEM to get enroll details the same time it is
unmasked. No change for Upstream enrollment or Settings page enrollment.

For Downstream enrollment, flow will be (assume exp enabled):

1) Card is to be unmasked, and before sending unmask request, if card is
eligible for VCN, send the InitVirtualCardEnroll call with no callback.

2) When form is submitted and a matched server card is extracted, invoke
the second InitVirtualCardEnroll, this time with same info but with a
callback to trigger the UI.

2.1) if response has been received at this time, invoke the callback
immediately to show the UI (virtual card enroll bubble/bottomsheet).

2.2) if response is not yet received from the server, assign the
callback, and OnDidGetDetailsForEnrollResponse will show the UI when
response is received.

Bug: 403617982
Change-Id: Iebd155fa42fbc6802452615609e3a3460897ae46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6638397
Reviewed-by: Olivia Saul <osaul@google.com>
Commit-Queue: Siyu An <siyua@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477567}