Update layer type used for the client view of a BubbleDialog
Currently the client view of BubbleDialog always paints to a textured
layer, however, implementations can specify the layer type of the
widget's layer.

So we can have NOT_DRAWN_LAYER for the widget but have a TEXTURED_LAYER
for the client view. This is unwanted because if the widget has a
NOT_DRAWN_LAYER, we expect the bubble to act as a transparent container.
Therefore it is unnecessary for client view to have a textured layer.

A NOT_DRAWN_LAYER is enough for the clipping purposes.

This CL updates the patter to configure the layer type from overriding a
method to having a setter method.

Bug: b:414655934, b:373875987, b:378126224
Change-Id: I4720720ba679bc2d4450450648dc81b3d96b69d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6655504
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Zoraiz Naeem <zoraiznaeem@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477571}