[B4B][a11y] Highlight focused folder in Bookmark Manager in HCM
When in high contrast mode (HCM), this creates a border around the
currently focused folder node in the tree view of the bookmark manager
so that it is visible which folder is currently selected. The selected
arrow color is set to the button text color so that it is visible as
well.

Screenshot: https://screenshot.googleplex.com/5vdSNk2tCeFZwxR
Screen recording:
https://screencast.googleplex.com/cast/NTUxODAyMTk4NzY2Mzg3MnxlOTk0ZWEwOC01ZA

Bug: 422480449
Change-Id: I246b8ef3407fa357e397017e05c17148b64a06d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6658433
Reviewed-by: Duong Dac <ddac@chromium.org>
Commit-Queue: Amelie Schneider <amelies@google.com>
Cr-Commit-Position: refs/heads/main@{#1477777}