Make AcceptCHFrameInterceptor handles added client hints
We aim to offload the AreCriticalCH handling from the browser
process's observer to the network layer if possible.
This CL implements the logic for checking that in
AcceptCHFrameInterceptor and is a part of series of CLs.

The feature works in the following way:
[1] Get client hints in the browser process on a request creation.
    https://chromium-review.googlesource.com/c/chromium/src/+/6638933
[2] Plumb the client hints to the network layer.
[3] Verify the client hints on connection time.
(You can see the proof of concept in https://chromium-review.googlesource.com/c/chromium/src/+/6630130)

To minimize the change size per CL, the CL is written in
[1] -> [3] -> [2], and this is the [3] CL.
That is because the mojo change requires full picture of the
change, and I believe both sides should be ready before its change.

Bug: 406407746
Change-Id: Ib651c165a03390ac3054be17feba28df05d3239b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6656395
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477806}