disk_cache: Support large file on PrefetchData
This CL lets SimpleSynchronousEntry::PrefetchData accepts uint64_t size
file. The data read at once must be smaller, but the total file size
could exceed size_t limit.

It also updates the trailer size to uint32_t from int32_t. The default
value was -1 and now 0. This is used for the current conditioning which
is done as `size <= 0` manner so 0 and -1 will be handled in the same
way. This is tested in the existing tests.

This is a preparation for large file support.

This CL alone does not affect the behavior so far since the file size
and offset is already wrapped to int range incorrectly before coming
into PrefetchData.
It will be fixed in the following cls.

Bug: 391398191
Change-Id: Id48e4e5d059db6e1aeaf17d763b51a4bb64e4d4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6660880
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Eriko Kurimoto <elkurin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477864}