[A11yPerformance] Use better containers for accessibility
This change contains many optimizations in one bang:

- Prefer Abseil flat associative containers over STL map/set. This not
  only switches from logarithmic to amortized constant time ops, but
  also reduces heap churn (each node in a map/set is a distinct
  allocation). This should also result in better memory patterns
  overall.
- Proactively reserve space in containers that are expected to hold a
  known number of items.
- OrderedSetContent is now a move-only type to avoid copies.
- Some map/set operations have been optimized (e.g., reduce the # of
  lookups).
- Use STL algorithms in some cases to walk containers.
- Replace uses of StringPrintf with base::StrCat -- it's phenomenally
  less expensive.
- Replace use of std::basic_string<>::operator+ with base::StrAppend.

A side effect of this change is that observers' methods that are
invoked repeatedly may now be invoked in a different
order. Previously, use of sorted containers meant that some methods
were invoked with ascending AXNodeIDs. This is no longer the case.

Pinpoint shows a reduction of ~25% in
BrowserAccessibilityManager::OnAccessibilityEvents for
very-many-nodes.html. The improvements are algorithmic, so pages with
more nodes will reap larger benefits.

AX-Relnotes: n/a.
Bug: 391797366, 40919326
Change-Id: I165c3d190408b4b73543cb5629b34e3b9b083336
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6633620
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Greg Thompson <grt@chromium.org>
Auto-Submit: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477875}