[Android] Fix SegFault in generation controller
If a focus change happens, the password generation controller resets the
generation_element_data_ and the active_frame_driver_. But it doesn't
remove the entry point to generation which means ShowBottomSheet may be
triggered on the reset objects. This seems to work mostly because the
reset objects are copied before use (which is probably unintentional).

These copies operate on undefined default values and thus cause the
segfault.

This CL checks whether the objects are reset and returns nullptr and
false as appropriate which prevents the segfault.

The case reproduced on a SDK 34 emulator on the IMDB registration page.

Bug: 426469437
Change-Id: I536b78ebfb4557a7559e899ef7c3a3c184cb03c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6655305
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Auto-Submit: Friedrich Hauser <friedrichh@chromium.org>
Commit-Queue: Friedrich Hauser <friedrichh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477904}