[FPF] Add browser-layer changes to communicate between component and WebKit.
Add a singleton to store the ContentRuleList data and inform observers
when new data is available.

Add a ScriptBlockingRuleApplierService and factory to generate the
former. The applier service is observing the ContentRuleList singleton
for new data and applies it to WebKit and the factory is tied to the
ProfileIOS.

Bug: 418218469
Change-Id: I5fe9c84f11b79996c1e38f2d4aecc723a25ee639
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6659762
Reviewed-by: Martin Pan-Verde <thesalsa@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Ron Sailer <rsailer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477908}