[DBSC] Split cookie rotation result histogram by a trigger
This CL adds an extra suffix to split
"Signin.BoundSessionCredentials.CookieRotationResult" and
"Signin.BoundSessionCredentials.CookieRotationTotalDuration" histograms.
The new suffixes capture a reason that triggered a cookie rotation
request.

The new histograms are recorded by BoundSessionRefreshCookieFetcherImpl
that receives an immutable trigger enum value via constructor. The
trigger value is computed by the BoundSessionCookieControllerImpl that
initiates the fetch.

Also, a little bit of extra plumbing is added to
BoundSessionCookieRefreshServiceImpl to pass an information about
whether a session is initialized from a disk data on startup or a brand
new session is just being registered.

Fixed: 427411410
Change-Id: I329b98adba93206c7d92e2506256c1ff81170982
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6662724
Reviewed-by: Ryan Sultanem <rsult@google.com>
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477917}