Actor: delay after type tool before page stability
Drop downs often wait until a user finishes typing to fetch and display
a list of available options. This can be done via, e.g. setTimeout
that's reset on every keystroke. This means that after type tool injects
a key event sequence, the effects of that may not be fully processed
until after such a timeout.

This CL changes the minimum delay time into an observation delay - that
is, a delay before the page stability monitor starts to watch the page.
Like the minimum delay, this ensures that page stability waits at least
as long as this delay. Unlike the minimum delay, if network requests are
still pending at the end of this delay - or the main thread is busy, the
page stability monitor will delay until those settle. i.e. This delay
expands the amount of time the monitor is waiting for the tool to
execute.

Additionally add some more detailed logging to the monitor to help
diagnose issues in the future.

Fixed: 425979083
Change-Id: If9eb7a56113c1f6b8f7f988a2b7b9598f5b9f32e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6662859
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477922}