[iOS]Manage Accounts Coordinator’s stop close the view
The `stop` method should usually close its own view controllers.

Before this CL, the view was dismissed either:
* in `closeSettings`, which was called when this view is opened in the settings,
* when the parent view is dismissed by its own parent view, in the case of the account-menu,

This second case is not the way coordinators are suppsoed to works,
and also stops working with the fix in crrev.com/c/6662560.

After this CL, we guarantee that the view controller is always
dismissed exactly once. Either by `closeSettings` if in the settings,
or by `stop` otherwise.

Bug: 425899803
Change-Id: I6dc5a01a6f3e2ca1be9523cceac04cf5beed4d15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6666860
Commit-Queue: Arthur Milchior <arthurmilchior@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477956}