[views-ax] Introduce ViewAccessibilityAXTreeSerializer
This CL creates a new AXTreeSerializer for ViewAccessibility. It will be
used to send incremental updates of the accessibility tree created by
the views hierarchy. Most of the logic is already implemented in
AXTreeSerializer. Tests specific to the usage we'll make with views will
be added in a follow-up CL, once we start serializing the tree.

Bug: 40672441
Change-Id: I411a12e113f44b7e7b78a5a151482a253daff33d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6658479
Reviewed-by: Javier Contreras <javiercon@microsoft.com>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1478018}